+++ ethers-providers/src/provider.rs
@@ -848,12 +848,13 @@ impl<P: JsonRpcClient> Middleware for Provider<P> {
         self.subscribe([logs, filter]).await
     }
 
-    async fn fee_history<T: Into<U256> + serde::Serialize + Send + Sync>(
+    async fn fee_history<T: Into<U256> + Send + Sync>(
         &self,
         block_count: T,
         last_block: BlockNumber,
         reward_percentiles: &[f64],
     ) -> Result<FeeHistory, Self::Error> {
+        let block_count = block_count.into();
         let last_block = utils::serialize(&last_block);
         let reward_percentiles = utils::serialize(&reward_percentiles);
 
@@ -868,7 +869,7 @@ impl<P: JsonRpcClient> Middleware for Provider<P> {
         .or(self
             .request(
                 "eth_feeHistory",
-                [utils::serialize(&block_count.into().as_u64()), last_block, reward_percentiles],
+                [utils::serialize(&block_count.as_u64()), last_block, reward_percentiles],
             )
             .await)
     }